pyproject.toml
[project]
name = "jobrunner"
version = "0.1.0"
description = "Multi-user batch job orchestration service with FastAPI and Celery"
readme = "README.md"
authors = [
    { name = "Memetelve", email = "memetelve@gmail.com" }
]
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.111",
    "uvicorn[standard]>=0.30",
    "celery>=5.4",
    "redis>=5.0",
    "sqlalchemy>=2.0",
    "pydantic-settings>=2.2",
    "python-jose[cryptography]>=3.3",
    "passlib[bcrypt]>=1.7",
    "bcrypt<5",
    "email-validator>=2.2",
    "asyncpg>=0.29",
    "aiosqlite>=0.20",
    "httpx>=0.27",
    "python-multipart>=0.0.9",
    "jinja2>=3.1"
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3",
    "pytest-asyncio>=0.23",
    "ruff>=0.6",
    "mypy>=1.11"
]

[project.scripts]
jobrunner = "jobrunner.cli:main"
jobrunner-api = "jobrunner.cli:run_api"
jobrunner-worker = "jobrunner.cli:run_worker"

[build-system]
requires = ["uv_build>=0.9.5,<0.10.0"]
build-backend = "uv_build"
